Redo all scraping
OOOH BOY! This is a big one. Instagram removed the graphql object from
being sent along with the initial page load and instead does it on
a subsequent sub page load. Well, that's an issue since we have to
actually have to look out for these and intercept them, which this does.

This is a bit flakier, but overall works well and all tests pass.

This also fixes an issue with relogging in a lot even if you're already
logged in.

To test:
`bundle install`
`rake test`