Adjust so that the request is quit when done
For some reason, certain calls to Instagram will just hang, especially
when moving between different scrapers (zori -> forki for instance).
This doesn't come up when testing this in isolation, but does when in
Hypatia. I think it's a race condition.

Anyways, to fix this, we just quit the session when we have the data we
want and then it starts up again afterwards.

This also removes a redundent site visit when logging in.